+++ packager/view/mainWindow.py
@@ -1,5 +1,6 @@
 import logging
 
+
 from tkinter import ttk
 from tkinter import *
 import webbrowser